refactor :: change category import location
